Get Data: mz low/high cutoffs to Double. Add List mz low/high cutoffs
For the existing mz low/high cutoffs, change from Float to Double.
Add List of mz low/high cutoffs to Get_ScanDataFromScanNumbers_Request

+++ Common_Get_Data_Webapp_Client_Connector_Library/src/main/java/org/yeastrc/spectral_storage/get_data_webapp/shared_server_client/webservice_request_response/main/Get_ScansDataFromRetentionTimeRange_Request.java
@@ -36,13 +36,13 @@ public class Get_ScansDataFromRetentionTimeRange_Request extends BaseGetDataWebs
 	 * If populated, do not return any peaks with mz below this cutoff.  
 	 */
 	@XmlAttribute // attribute name is property name
-	private Float mzLowCutoff;
+	private Double mzLowCutoff;
 
 	/**
 	 * If populated, do not return any peaks with mz above this cutoff.  
 	 */
 	@XmlAttribute // attribute name is property name
-	private Float mzHighCutoff;
+	private Double mzHighCutoff;
 	
 	/**
 	 * Only return scans for this scan level
@@ -71,28 +71,28 @@ public class Get_ScansDataFromRetentionTimeRange_Request extends BaseGetDataWebs
 	/**
 	 * If populated, do not return any peaks with mz below this cutoff.  
 	 */
-	public Float getMzLowCutoff() {
+	public Double getMzLowCutoff() {
 		return mzLowCutoff;
 	}
 
 	/**
 	 * If populated, do not return any peaks with mz below this cutoff.  
 	 */
-	public void setMzLowCutoff(Float mzLowCutoff) {
+	public void setMzLowCutoff(Double mzLowCutoff) {
 		this.mzLowCutoff = mzLowCutoff;
 	}
 
 	/**
 	 * If populated, do not return any peaks with mz above this cutoff.  
 	 */
-	public Float getMzHighCutoff() {
+	public Double getMzHighCutoff() {
 		return mzHighCutoff;
 	}
 
 	/**
 	 * If populated, do not return any peaks with mz above this cutoff.  
 	 */
-	public void setMzHighCutoff(Float mzHighCutoff) {
+	public void setMzHighCutoff(Double mzHighCutoff) {
 		this.mzHighCutoff = mzHighCutoff;
 	}
